[BPK-4390]: Migrate components L-P to new format
