+++ packages/bpk-component-mobile-scroll-container/examples.css
@@ -1,0 +1,1 @@
+@keyframes bpk-keyframe-spin{100%{transform:rotate(1turn)}}.bpk-stories-mobile-scroll-container__leading-indicator::before{position:absolute;top:0;bottom:0;content:' ';display:block;z-index:1;width:2.25rem;background-image:linear-gradient(90deg, #0770e3, rgba(7,112,227,0));pointer-events:none;left:0}.bpk-stories-mobile-scroll-container__trailing-indicator::after{position:absolute;top:0;bottom:0;content:' ';display:block;z-index:1;width:2.25rem;background-image:linear-gradient(270deg, #0770e3, rgba(7,112,227,0));pointer-events:none;right:0}.bpk-scroll-block{display:flex}.bpk-scroll-block__item{display:flex;min-width:13.125rem;height:13.125rem;justify-content:center;align-items:center;background-color:#f1f2f8}.bpk-scroll-block__item--alternate{background-color:#dddde5}
