+++ packages/bpk-component-nudger/BpkNudgerStory.css
@@ -1,1 +1,1 @@
-@keyframes bpk-keyframe-spin{100%{transform:rotate(1turn)}}.bpk-nudger-outline{padding:.375rem;background-color:#0770e3}.bpk-nudger-configurable{width:7rem}
+@keyframes bpk-keyframe-spin{100%{transform:rotate(1turn)}}.bpk-nudger-outline{color:#fff}.bpk-nudger-configurable{width:7rem}
