+++ packages/bpk-component-nudger/BpkNudgerStory.scss
@@ -19,8 +19,7 @@
 @import '~bpk-mixins/index';
 
 .bpk-nudger-outline {
-  padding: $bpk-spacing-xs;
-  background-color: $bpk-color-sky-blue;
+  color: $bpk-color-white;
 }
 
 .bpk-nudger-configurable {
