+++ packages/bpk-component-nudger/stories.js
@@ -18,157 +18,21 @@
 
 /* @flow strict */
 
-import React, { Component } from 'react';
 import { storiesOf } from '@storybook/react';
-import { action } from 'bpk-storybook-utils';
-import BpkLabel from 'bpk-component-label';
-import { cssModules } from 'bpk-react-utils';
 
-import STYLES from './BpkNudgerStory.scss';
-
-import BpkNudger, { BpkConfigurableNudger } from './index';
-
-const getClassName = cssModules(STYLES);
-
-class NudgerContainer extends Component<{}, { value: number }> {
-  constructor() {
-    super();
-
-    this.state = {
-      value: 2,
-    };
-  }
-
-  handleChange = value => {
-    this.setState({ value });
-  };
-
-  render() {
-    return (
-      <div>
-        <BpkLabel htmlFor="my-nudger">Number of passengers</BpkLabel>
-        <BpkNudger
-          id="my-nudger"
-          min={1}
-          max={10}
-          value={this.state.value}
-          onChange={this.handleChange}
-          decreaseButtonLabel="Decrease"
-          increaseButtonLabel="Increase"
-        />
-      </div>
-    );
-  }
-}
-
-const options = ['economy', 'premium', 'business', 'first'];
-const compareValues = (a: string, b: string): number => {
-  const [aIndex, bIndex] = [options.indexOf(a), options.indexOf(b)];
-  return aIndex - bIndex;
-};
-
-const incrementValue = (a: string): string => {
-  const currentIndex = options.indexOf(a);
-  const newIndex = currentIndex + 1;
-  if (currentIndex === -1 || newIndex >= options.length) {
-    return a;
-  }
-  return options[newIndex];
-};
-
-const decrementValue = (a: string): string => {
-  const index = options.indexOf(a) - 1;
-  if (index < 0) {
-    return a;
-  }
-  return options[index];
-};
-
-const formatValue = (a: string): string => a.toString();
-
-class ConfigurableNudgerContainer extends Component<{}, { value: string }> {
-  constructor() {
-    super();
-
-    this.state = {
-      value: 'premium',
-    };
-  }
-
-  handleChange = value => {
-    this.setState({ value });
-  };
-
-  render() {
-    return (
-      <div>
-        <BpkLabel htmlFor="nudger">Traveller Class</BpkLabel>
-        <BpkConfigurableNudger
-          id="nudger"
-          min="economy"
-          max="first"
-          value={this.state.value}
-          onChange={this.handleChange}
-          decreaseButtonLabel="Decrease"
-          increaseButtonLabel="Increase"
-          compareValues={compareValues}
-          incrementValue={incrementValue}
-          decrementValue={decrementValue}
-          formatValue={formatValue}
-          inputClassName={getClassName('bpk-nudger-configurable')}
-        />
-      </div>
-    );
-  }
-}
+import {
+  DefaultExample,
+  LowerBoundExample,
+  UpperBoundsExample,
+  StatefulExample,
+  ConfigurableExample,
+  OutlineExample,
+} from './examples';
 
 storiesOf('bpk-component-nudger', module)
-  .add('Default', () => (
-    <BpkNudger
-      id="my-nudger"
-      min={1}
-      max={99}
-      value={3}
-      onChange={action('change')}
-      decreaseButtonLabel="Decrease"
-      increaseButtonLabel="Increase"
-    />
-  ))
-  .add('Lower bounds', () => (
-    <BpkNudger
-      id="my-nudger"
-      min={3}
-      max={99}
-      value={3}
-      onChange={action('change')}
-      decreaseButtonLabel="Decrease"
-      increaseButtonLabel="Increase"
-    />
-  ))
-  .add('Upper bounds', () => (
-    <BpkNudger
-      id="my-nudger"
-      min={1}
-      max={99}
-      value={99}
-      onChange={action('change')}
-      decreaseButtonLabel="Decrease"
-      increaseButtonLabel="Increase"
-    />
-  ))
-  .add('Stateful example', () => <NudgerContainer />)
-  .add('Configurable nudger', () => <ConfigurableNudgerContainer />)
-  .add('Outline nudger', () => (
-    <div className={getClassName('bpk-nudger-outline')}>
-      <BpkNudger
-        id="outline-nudger"
-        min={1}
-        max={99}
-        value={3}
-        onChange={action('change')}
-        decreaseButtonLabel="Decrease"
-        increaseButtonLabel="Increase"
-        buttonType="outline"
-      />
-    </div>
-  ));
+  .add('Default', DefaultExample)
+  .add('Lower bounds', LowerBoundExample)
+  .add('Upper bounds', UpperBoundsExample)
+  .add('Stateful example', StatefulExample)
+  .add('Configurable nudger', ConfigurableExample)
+  .add('Outline nudger', OutlineExample);
