+++ packages/bpk-component-panel/examples.js
@@ -1,0 +1,50 @@
+/*
+ * Backpack - Skyscanner's Design System
+ *
+ * Copyright 2016-2021 Skyscanner Ltd
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import React from 'react';
+
+import BpkPanel from './index';
+
+const DefaultExample = () => (
+  <BpkPanel>
+    Lorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque
+    imperdiet lobortis tellus, non rhoncus erat tincidunt id. Pellentesque
+    consectetur, dolor nec vulputate vehicula, ex metus mattis ante, non dictum
+    mi ante eu arcu.
+  </BpkPanel>
+);
+
+const WithoutPaddingExample = () => (
+  <BpkPanel padded={false}>
+    Lorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque
+    imperdiet lobortis tellus, non rhoncus erat tincidunt id. Pellentesque
+    consectetur, dolor nec vulputate vehicula, ex metus mattis ante, non dictum
+    mi ante eu arcu.
+  </BpkPanel>
+);
+
+const FullWidthExample = () => (
+  <BpkPanel fullWidth>
+    Lorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque
+    imperdiet lobortis tellus, non rhoncus erat tincidunt id. Pellentesque
+    consectetur, dolor nec vulputate vehicula, ex metus mattis ante, non dictum
+    mi ante eu arcu.
+  </BpkPanel>
+);
+
+export { DefaultExample, WithoutPaddingExample, FullWidthExample };
