+++ packages/bpk-component-paragraph/examples.js
@@ -1,0 +1,46 @@
+/*
+ * Backpack - Skyscanner's Design System
+ *
+ * Copyright 2016-2021 Skyscanner Ltd
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/* @flow strict */
+
+import React from 'react';
+
+import BpkParagraph from './index';
+
+const DefaultExample = () => (
+  <div>
+    <BpkParagraph>
+      Lorem ipsum dolor sit amet, consectetuer adipiscing elit. Aenean commodo
+      ligula eget dolor. Aenean massa. Cum sociis natoque penatibus et magnis
+      dis parturient montes, nascetur ridiculus mus.
+    </BpkParagraph>
+    <BpkParagraph>
+      Donec quam felis, ultricies nec, pellentesque eu, pretium quis, sem. Nulla
+      consequat massa quis enim. Donec pede justo, fringilla vel, aliquet nec,
+      vulputate eget, arcu. In enim justo, rhoncus ut, imperdiet a, venenatis
+      vitae, justo.
+    </BpkParagraph>
+    <BpkParagraph>
+      Nullam dictum felis eu pede mollis pretium. Integer tincidunt. Cras
+      dapibus. Vivamus elementum semper nisi. Aenean vulputate eleifend tellus.
+      Aenean leo ligula, porttitor eu, consequat vitae, eleifend ac, enim.
+    </BpkParagraph>
+  </div>
+);
+
+export default DefaultExample;
