+++ packages/bpk-component-progress/examples.css
@@ -1,0 +1,1 @@
+.bpkdocs-progress-container__flex{display:flex;justify-content:space-between;align-items:center}.bpkdocs-progress-container__flex-grow{flex-grow:1}
