+++ packages/bpk-component-progress/examples.scss
@@ -1,7 +1,7 @@
 /*
  * Backpack - Skyscanner's Design System
  *
- * Copyright 2016-2021 Skyscanner Ltd
+ * Copyright 2016-2019 Skyscanner Ltd
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,14 +16,14 @@
  * limitations under the License.
  */
 
-@import '~bpk-mixins';
+.bpkdocs-progress-container {
+  &__flex {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
 
-.bpk-stories-mobile-scroll-container {
-  &__leading-indicator {
-    @include bpk-scroll-indicator-left($bpk-color-sky-blue);
-  }
-
-  &__trailing-indicator {
-    @include bpk-scroll-indicator-right($bpk-color-sky-blue);
+    &-grow {
+      flex-grow: 1;
+    }
   }
 }
